+++ lib/dunolint/src/dune_project0/implicit_transitive_deps.mli
@@ -19,6 +19,8 @@
 (*_  <http://www.gnu.org/licenses/> and <https://spdx.org>, respectively.         *)
 (*_********************************************************************************)
 
+module Value = Implicit_transitive_deps__value
+
 module Predicate : sig
-  type t = [ `equals of bool ] [@@deriving compare, equal, sexp]
+  type t = [ `equals of Value.t ] [@@deriving compare, equal, sexp]
 end